This project is going to recreate the Todo application found in chapter 4 of Backbone Fundamentals using Uncle Bob's Three Laws of Test Driven Design.  The Three Laws of TDD are
1. You are not allowed to write any production code unless it is to make a failing test pass.
2.  You are not allowed to write any more of a unit test than is sufficient to fail.
3. You	are not allowed to write more production code than is sufficient to make one failing unit test pass.

Following these laws, this means that we must first set up a testing framework before writing a single line of code.  Since we are using javascript, we are setting up the javascript testing framework jasmine.

Jasmine can work as a standalone as long as you have a browser.  In order to run jasmine, open the html file SpecRunner.  It will list out hat tests are running and if they passed.