In this commit, we wrote a single test to see if a mdel for our todo list item existed.  the test fails because we have not written any source code yet.
Why was this test written this way?  In chapter 4 of Backbone Fundamentals, there is a list of features tht need to be implemented.  The first feature is a todo model to describe individual todo list items.  since these items are the core of the application, it makes sense to implement the model first.

Which begs the question: what is a model and how to we test for it? Accoring to Backbone fundamentals Chapter Three, a model is a container for data.  I personally like to think of it like a model airplane kit- like a model airplane kit, a model is a box that contains a bunch of useful and related parts in an organized way so tht you can do something interesting with them later.

Here's another tricky part: how do you make a model?  Well, turns out my idea of a model was incorrect- it is not like a model airplane kit.  Instead, a model is more like the factory that makes model airplane kits.  See, usually the reason you have a model in the irst place is o be able to quickly create a particular type of container for a bit of data.  Take our todo list item.  The the phrase todo list item is like a container:  you know that whatever is in the item is something you need to do.  And it's the model's job to create those containers.

Without a model, you can't make todo items or model airplane boxes or hatever sort of ontainer your model is set up to make.  So it seems reasonable that first thing to test for is whether a model exists.

Now it's time to dig into how backbone creates a model.  Looking at Chapter 3in backbone fundamentals, it seems like it creates a model uing the magic words Backbone.Model.extend({});, and set that whole thing equal to a variable.  So it makes sense to test the variable that the model was assigned to, and see if it has been defined.  If it has been defined, that means that the magic words worked.

(I guess, technically, I really ought to be checking what is in the variable is actually a model, but I don't want to overly complicte my learning process.