So I changed the test.  instead of just checking if The Todo variable is defined, I check to see if it is a backbone model.  I do this by calling the backbone create model function, and run it.  The test fails.
The way it failes tels us something interesting.  The variable backbone is not defined.  Now, backbone is supposed to be defined by the backbone library.  So, those of you who have been paying attention might have noticed that I never got around to includeing the backbone library in my source code.  This is because I am trying to follow  the wisdom of tdd, which is to never write a line of production code unless it is to make a failing test pass.  So now that our tests have shown that we are missing a library, it is time to include it.