So, I've been trying to test whether the ing I get back from backbone when I call Backbone.Model.extend({}) is a model.  The thing is, at this point I don't really know what a model is.  I'm using jasmine and TDD to teach myself backbone.  I probably won't know how to test for what something is until I am a lot  more familiar with using Backbone and Jasmine.  So it's not at all useful for me to write a test for something I don't know how to test.
What is useful is to write a test to see if calls to the library are actuslly working.  The most basic thing I need to know about a library is if I make a call to the library, does something happen?  I something happens, the library is working.  If nothing happens, the library is probably missing.

As my broth would say, the only way to test if a library works is to try and use it.  The two tests added try to do things that work if the library is present and break if the library is not.