Turns out tht backbone views have a hard dependency on jquery.  if you don't have jquery loaded, the backbone views constructor will not work,  because I think backbone by default asumes you are using jquery to manipulate Dom objects.
Also, some thoughts on views: a view's job is to say what a piece of your applicaion will look like.  Think of your model as a fasion model and the view is the clothing the fashion model wears.  A view needs to manage two things:  it needs o go and get data from the model, and it needs to dress that data up in a pretty outfit so it can go to the dance.  Now, usually the reason you have a view is because you have a particular (uually rectangular) box on your webpage that iu want o fill with data. That box is annelement.  So really, a view takes data from a model, dresses it up, and stick it in a particulr box (aka element ) on the webpage.  so it needs to know where the model is, how to dress the model's data, and where the box that it is going to stick the final result in.

which means those are the three things we need to figure out how to test.