This was a tough one o decide what to test next.  It took me awhile to begin wrappng my head around how to test views.  Basically, it seems to be boiling down to this:  a view has a box called el.  El is like  a picture frame.  the view has a function called render.  render paints the picture you stick in the picture frame.
you test it by looking inside the picture frame at the picture and seeing if you gt what you wanted.  since the entire 'picture' is 'painted' in html, you just compare the html the render put in el to the html you expected, and see if they are correct.

what was weirding me out about testing views is that I thought I would have to render the html to an actual page and physically look at it to see if it had what I wanted, and that did not seem very scalable at all.