So far the most difficult thing for me has been to build only what I need to build.
There are so many things that I don't need to be doing in order to learn.  For example, I don't need to set the El property, and I don't need to know how to set the el property, because the default div is just fine or sticking in some code.  And right now, I really don't need to know about templates because they don't make my life easier at this point.

Anyway, I do need to know how to output a checked checkbox to reflect the data in the todo model. Sotht's what I did. Yay two new tests.