After two days of hacking, I finally figured out why my event was not binding. you can't, apparently, bind an input to a click event.  you have o bind an input to a change event.  you can bind a paragraph to a change event.
i also moved the setup of my test model and and view to beforeEach in testrunner and got rid of a unch of distracting tests.

this was the most frustrating problem  have come across so far.  I came up with several theories.  I still have no good theory as to why the previous commit's test passed when you passed in a list element with li in it, or why if I passed in a div into the el it replaced my whole jasmie page with the contents of el.

things I learned:  there is a buttload left to learn.  I would have diagnosed the prblem much sooner if I had changed the input tag to a p tag, but I didn't want to break all of the tests hat rely on outputting specific xml.  in short, you really want the solution to be omething tht doesn't requre a lot of work, or fixing somethng that you know will be tedious. But that also implies hat I ral
hould stop munging together html strings and start using templates that allow me to change my tests and

easily.