now it is time to clean up the mess that is my tooview
render function.  Right now I am munging html strings together into the content of my todoview nd shoving it in the el property.  I am mixing presentation with content. So I am going to ttempt in the next few commits, to step by step untangle the munge and put it in a proper templating system.

now, ideally, if you have a template, all you want to do is pass it data and have it magically output a pretty presentation of the data.  let's work on that.