Set collection so that view updates whenever a model is added.
point of order: apparently collections do not have change events.  This is a deliberatedesign philosophy of backbone, because collections are arrays of models and apparently yiu can only add things, remove things. and reset things in a collection, so there is no need for a change event because there are so few possible events in the first place.