Added an input box to the Todo template for editing the caption.  Added a function to get value from edit box to update the caption.  The function is called changeCaption.
This took me awhile to wrap my head around, because I was thinking about ways to replace the label with the edit bx and vice versa.  It would have resulted in really messy code.  It is much easier just to have all of the elements yiu need in a view and use css to hide or show them as necessary.