Tested if view's clear event is called when delete button is clicked
What I did:  added a test to see if clear event was called when delete button was clicked. To make test pass, added click event to events property in the view.  Added delete button to view template in Specrunner.

What I learned:  when spying on an object in jasmine, if the method you are spying on is triggered by an event, you need to call the delegateEvents() function on the view so that the event will be bound to the spy and not the actual method.  Also leaned how to use spies in Jasmine