Tested to see if deleting an item rerenders the todo list view.
what I did: wrote a test to see if removing an item changed the collection view's html, inficating that it rerendered.  in the test, the item showes up in todo_list_view's el property. and after the delete button is clicked, it disappears from the todo_list_view's el property, indicating that the item's html has been removed from the dom.  i

what I learned:  The fact that this test passes was unexpected to me, because when I added an event to the collection, I had to have the collection's view listen for the collection's add event in order to call the collection's render function so that the el property has the most up-to-date information.  However, I did not have to listen for a delete event in the collection view in order to update the el property.  I theorize that this is because when you click on the delete button, that event is handled by the todo item view, and the function, clear, that it calls, removes the item from dom tree.  Since part of the Dom tree is contained inside the collection view's el property, when a todo item removes itself from the dom tree, it also removes itself from the collection view's el as well.  That means the collection view does not have to listen for deletes because the delete is handled at the item level.