Added index page and init function so that I can actually see my app in action.
What I did:  Wrote an index page and and initializer to so that I could actually see app in action.

What I learned:  the checkbox toggles whenever any input box changes, not just when the checkbox is clicked.  Doubleclick on the label is not functioning.  Other than that, MY CREATION LIVES!  ARISE, FRANKENSTEIN!  WOOOoOOOoOooOOO!