Fixed toggle checkox bug and label doubleclick bug
What I did:  Fixed two bugs.  Bug number one was the fact tht the checkbox toggled on and off whenever any input was changes in a odo item.  So if you edited the caption, you also checked or unchecked the chekbox.  This was because the view was listening for a change in any input, not just a change in the checkbx input element.  Fixed by gving checkbo input a classname "toggle" and listening for changes to the element with the toggle classname.
The second bug was the fact that when I doubleclicked on a label,the the fous was not switching to the edit box.This is probably because I am developing on an ipad and the doubletap event already has another specific meaning in mobile browsers.  Instead of trying o get oubletap to work, I changed it to a single click event and it works just fine.

What I learned:  Tests are great, but sometimes you don't find bugs until you actially use the app.  On the other hand, with all the teting I did, the bugs were easy to find and fix. Questions for the future include whether I should have been able to write tests that would have caught those bugs in the first place.