Added a completedFilter function to todoListView to identify todos that have been completed.
The completedFilter's job is to take in a todo and check if it has been completed.  If it has, completed-filter will return it si that it can either be rendered or passed through some other filter.  if it is not completed, the function returns null.  Basically it filters out todos that are not completed.

What I learned:  Finding the smallest possible step to implement a feature is hard.  Really hard.  I spent all morning deciding  figuring out that the first thing I needed to do was have a way to choose which todos I wanted to render.  So it makes sense to dump it through a function and see what comes out.