Edited render function to add a filter based on flags
What I did:  edited thre render function so that if the show_only_completed flag is set, it would only bother to render the todos that are complete.

What I learned:  This is a messy addition to the render function. It is full of state.  I'm ure there is a better way to do this.