Added a toggleShowOnlyCompleted function that toggles the showOnlyCompleted flag on and off.
What I did:  This toggles the flag on and off, and the render function uses the flag to determine what to render. it also triggers the render function so that the only todos that are shwn are the ones we want.

What I learned: I probably should not have made this a toggle.  I do not feel good about any of the code I have written today.