changed toggleShowOnlyCompleted to turnOnShowOnlyCompleted, and had it only set show_only_completed flag to true
What I did:  changed the toggle to a turn on because this function will be called when we push a button to turn it on.  toggle is not needed at this point.

what I learned:  Right now, I'm writing tests by working backwards.  I'm working form the change in the render function out to the button, making sure every link in the chain works, and writing a test for each link in the chain.  Trouble is, my nose is wrinkling, and I think most of what I've done to build this feature is a code smell.  Ick.