Tried to add classname to todoListView to get it to have a clss for css to select.
What I did:  Added a className property to todoLitView.  The todoListView constructor uses these properties to construct the containing tag for the todoListView's el.

What I learned:  This did not work as intended because in init.js I passed in an existing tag, to todoListView, which overwrite he default tags.