I attempted to add css to make the input boxes or editing the todo items hidden until one clicked on the label of the todo they wanted to edit.
What I did: The following did not work:  When you click on the caption of a todo item, the page focuses on the editbox right next to it- i.e. it programatticalle sets the focus on that todo item's editbox.  I wanted to hide all of the boxes that didn't have focus, so I set the visibility of .edibox to hidden.  Since clicking on a caption set the focus on the associated editbox, I tried to say if .edibox had focus, make it visible.  The trouble is, this doesn't work because you can't set the focus on something that is hidden.

what I learned:  the browser does not let you programmatically focus on an element that is hidden, so this technique did not work and in fact broke the program.