Added and tested  an editing state: a css class that appears when a todo item is being edited and diappears when the item is finished editing.
What I did:  Added two lines of code: one to the edit function, that sets a class on the todo item called edit.  This is so the css can display the proper editing stylesheet.  The class is removed when the changeCaption function is called, because it corrosponds to the end of the editing session.  I modified two tests to test for these changes, and added css that hides the editbox when the todo item does not have the editing class and shows the editbox and hides the caption when the todo item has the eiting class.

What I learned:  I was basically setting it up so that the todo item had an editing state.  I don't beleive I like state, because it made two of my functions violate the "only do one thing" principle, and makes it sof I have to remember what state things are in before I call a function.  It makes things messy.r